Refactored infrastructure into single file
