+++ test/data/mssql/integration.dynamicSchema.tests.js
@@ -1,7 +1,7 @@
 // ----------------------------------------------------------------------------
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // ----------------------------------------------------------------------------
-var config = require('../../infrastructure/configuration')().data,
+var config = require('../../appFactory').configuration().data,
     dynamicSchema = require('../../../src/data/mssql/dynamicSchema'),
     statements = require('../../../src/data/mssql/statements'),
     execute = require('../../../src/data/mssql/execute'),
