+++ test/data/mssql/integration.tests.js
@@ -4,7 +4,7 @@
 var index = require('../../../src/data/mssql'),
     execute = require('../../../src/data/mssql/execute'),
     queries = require('../../../src/query'),
-    config = require('../../infrastructure/configuration')().data,
+    config = require('../../appFactory').configuration().data,
     expect = require('chai')
         .use(require('chai-subset'))
         .use(require('chai-as-promised'))
