+++ test/express/integration/notifications.tests.js
@@ -8,7 +8,7 @@ var sinon = require('sinon'),
         .expect,
     request = require('supertest-as-promised'),
     express = require('express'),
-    mobileApps = require('../../infrastructure/appFactory').ignoreEnvironment,
+    mobileApps = require('../../appFactory').ignoreEnvironment,
     nhStub, notifFactoryStub, app, installation;
